requirements.txt
paho-mqtt
pymodbus
msgpack